% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_points.R
\name{calc_points}
\alias{calc_points}
\title{Calculate points in food}
\usage{
calc_points(food_id, quantity = 100, measure = "g", density = 1,
  point_system = "wwpp", decimals = 0)
}
\arguments{
\item{food_id}{\code{character} Food ID in database.}

\item{quantity}{\code{numeric(1)} Quantity of food. Default is 100.}

\item{measure}{\code{character(1)} Measure to use for food. Available measures are:
\itemize{
  \item \code{'g'} grams (default)
  \item \code{'kg'} kilograms
  \item \code{'oz'} ounces
  \item \code{'lb'} pounds
  \item \code{'l'} liters
  \item \code{'ml'} mililiters
  \item \code{'c'} cups
  \item \code{'pt'} pints
  \item \code{'qt'} quarts
  \item \code{'gal'} gallons
}}

\item{density}{\code{numeric(1)} Density to use for volume-to-mass
conversions in grams per mililiter. Default is 1 g / ml, or the density of
water.}

\item{point_system}{\code{character} Points system to use: \code{'wwp'} for
Weight Watchers Points, \code{'wwpp'} for Weight Watchers Points Plus. Default
is \code{'wwpp'}.}

\item{decimals}{\code{numeric(1)} Number of decimal places to show. Defaults
is 0.}
}
\description{
Calculates points of a given quantity of food
}
