% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_flat_file.R
\name{display_flat_file}
\alias{display_flat_file}
\title{Display flat file}
\usage{
display_flat_file()
}
\description{
Creates a flat file with all the information in the food database for easy reference
}
