% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_food.R
\name{find_food}
\alias{find_food}
\title{Find food in database}
\usage{
find_food(food, return_type = "all", case_sens = FALSE)
}
\arguments{
\item{food}{\code{character(1)} Food name to search for}

\item{return_type}{\code{character(1)} What to return? Options are:
\itemize{
  \item \code{'all'} Return all matches (default)
  \item \code{'first'} Return first match
}}

\item{case_sens}{\code{logical(1)} Case sensitive search? Default is FALSE}
}
\description{
Given a character string, search for a match in the database
}
