% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrition_data.R
\name{get_nutrition_data}
\alias{get_nutrition_data}
\title{Get nutrition data from database}
\usage{
get_nutrition_data(food_id)
}
\arguments{
\item{food_id}{\code{integer(1)} Food ID number (NDB_No)}
}
\description{
Given a food type and amount, provides nutrition information on that food
}
