% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wwp_data.R
\name{get_wwp_data}
\alias{get_wwp_data}
\title{Get Weight Watchers Points data (Original)}
\usage{
get_wwp_data(food_id)
}
\arguments{
\item{nutrition_data}{\code{data.frame} A food ID from the database}
}
\description{
Get data required to calculate points on Weight Watchers Points (Original)
system. See:
\url{http://calorielab.com/news/2015/10/08/weight-watchers-pointsplus-formula/}
}
