% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_food_names.R
\name{list_food_names}
\alias{list_food_names}
\title{List food names.}
\usage{
list_food_names(desc_type = "short")
}
\arguments{
\item{desc_type}{\code{character(1)} Description type to use, 'long' or 'short'. Default is 'short'.}
}
\description{
Create a vector with the names of all avialable foods, to be used for partial matches
}
