% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwp.R
\name{wwp}
\alias{wwp}
\title{Calculate Weight Watchers Points (original)}
\usage{
wwp(food_id = NULL, wwp_data = NULL, calories = NULL, fat = NULL,
  fiber = NULL)
}
\arguments{
\item{food_id}{\code{character(1)} A food ID from the database}

\item{wwp_data}{\code{numeric} A vector of numbers with required data from \link{get_wwp_data}}

\item{calories}{\code{numeric(1)} calories in food (KCal)}

\item{fat}{\code{numeric(1)} grams of fat}

\item{fiber}{\code{numeric(1)} grams of fiber}
}
\description{
Calculate Weight Watchers Points (original) from a food_id, the result of
get_wwp_data, or a set of inputs. See:
\url{http://calorielab.com/news/2015/10/08/weight-watchers-pointsplus-formula/}
}
