% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwpp.R
\name{wwpp}
\alias{wwpp}
\title{Calculate Weight Watchers Points Plus}
\usage{
wwpp(food_id = NULL, quantity = 1, wwpp_data = NULL, protein = NULL,
  carbs = NULL, fat = NULL, fiber = NULL)
}
\arguments{
\item{food_id}{\code{character(1)} A food ID from the database}

\item{wwpp_data}{\code{numeric} A vector of numbers with required data from \link{get_wwpp_data}}

\item{protein}{\code{numeric(1)} grams of protein}

\item{carbs}{\code{numeric(1)} grams of carbs}

\item{fat}{\code{numeric(1)} grams of fat}

\item{fiber}{\code{numeric(1)} grams of fiber}
}
\description{
Calculate Weight Watchers Points Plus from a food_id, the result of
get_wwpp_data, or a set of inputs. See:
\url{http://calorielab.com/news/2015/10/08/weight-watchers-pointsplus-formula/}
}
